1. Add strings in R.string 2. Change member data type of PlayList from int to long to store song ID 3. Add two methods in MainActivity for loading songs from storage
Note: Still have errors to correct in MainActivity

+++ app/src/main/java/com/example/android/musicplayer/PlayList.java
@@ -1,15 +1,15 @@
 package com.example.android.musicplayer;
 
 public class PlayList {
-    private int mMedia;
+    private long mMedia;
     private String mSongName;
 
-    public PlayList(int mediaSource, String songName){
+    public PlayList(long mediaSource, String songName){
         this.mMedia = mediaSource;
         this.mSongName = songName;
     }
 
-    public int getmMedia(){
+    public long getmMedia(){
         return this.mMedia;
     }
